Added ipres2018 Code of Conduct
added the link to the 2018 code of conduct. also commented on ipres entry for "organizations with unkonwn status". while there is a code of conduct for ipres 2018 in place now, this was written / put into place by the 2018 organizing team. there is currently no overaching iPRES conference series code of conduct yet.